Split Widgets into seperate files / modules
This will hopefully make them easier to maintain in the long term. The main polycore config all the examples and tests have been updated although the test suite isn't rendering for me. For now the split is as follows, but open to changes and suggestions:
  * widget_core * Widget * Rows * Columns * Filler * Frame
  * widget_cpu
    * Cpu
    * CpuRound * CpuFrequencies
  * widget_drive * Drive
  * widget_gpu
    * Gpu
    * GpuTop
  * widget_graph - This could have a better name * Bar * Graph * LED
  * widget_memory
    * MemoryBar
    * MemoryGrid
  * widget_network * Network
  * widget_text
    * Text
    * StaticText * TextLine
