+++ docs/modules/polycore.html
@@ -44,7 +44,14 @@
   <li><a href="../modules/data.html">data</a></li>
   <li><strong>polycore</strong></li>
   <li><a href="../modules/util.html">util</a></li>
-  <li><a href="../modules/widget.html">widget</a></li>
+  <li><a href="../modules/widget_core.html">widget_core</a></li>
+  <li><a href="../modules/widget_cpu.html">widget_cpu</a></li>
+  <li><a href="../modules/widget_drive.html">widget_drive</a></li>
+  <li><a href="../modules/widget_gpu.html">widget_gpu</a></li>
+  <li><a href="../modules/widget_graph.html">widget_graph</a></li>
+  <li><a href="../modules/widget_memory.html">widget_memory</a></li>
+  <li><a href="../modules/widget_network.html">widget_network</a></li>
+  <li><a href="../modules/widget_text.html">widget_text</a></li>
 </ul>
 
 </div>
@@ -245,7 +252,7 @@
 </div> <!-- id="main" -->
 <div id="about">
 <i>generated by <a href="http://github.com/stevedonovan/LDoc">LDoc 1.4.6</a></i>
-<i style="float:right;">Last updated 2023-03-26 23:57:51 </i>
+<i style="float:right;">Last updated 2023-05-01 20:49:18 </i>
 </div> <!-- id="about" -->
 </div> <!-- id="container" -->
 </body>
