+++ docs/modules/widget_gpu.html
@@ -1,0 +1,159 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"
+   "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
+<html>
+<meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
+<head>
+    <title>Reference</title>
+    <link rel="stylesheet" href="../ldoc.css" type="text/css" />
+</head>
+<body>
+
+<div id="container">
+
+<div id="product">
+	<div id="product_logo"></div>
+	<div id="product_name"><big><b></b></big></div>
+	<div id="product_description"></div>
+</div> <!-- id="product" -->
+
+
+<div id="main">
+
+
+<!-- Menu -->
+
+<div id="navigation">
+<br/>
+<h1>polycore</h1>
+
+<ul>
+  <li><a href="../index.html">Index</a></li>
+</ul>
+
+<h2>Contents</h2>
+<ul>
+<li><a href="#Class_Gpu">Class Gpu </a></li>
+<li><a href="#Class_GpuTop">Class GpuTop </a></li>
+</ul>
+
+
+<h2>Modules</h2>
+<ul class="nowrap">
+  <li><a href="../modules/cairo_helpers.html">cairo_helpers</a></li>
+  <li><a href="../modules/data.html">data</a></li>
+  <li><a href="../modules/polycore.html">polycore</a></li>
+  <li><a href="../modules/util.html">util</a></li>
+  <li><a href="../modules/widget_core.html">widget_core</a></li>
+  <li><a href="../modules/widget_cpu.html">widget_cpu</a></li>
+  <li><a href="../modules/widget_drive.html">widget_drive</a></li>
+  <li><strong>widget_gpu</strong></li>
+  <li><a href="../modules/widget_graph.html">widget_graph</a></li>
+  <li><a href="../modules/widget_memory.html">widget_memory</a></li>
+  <li><a href="../modules/widget_network.html">widget_network</a></li>
+  <li><a href="../modules/widget_text.html">widget_text</a></li>
+</ul>
+
+</div>
+
+<div id="content">
+
+<h1>Module <code>widget_gpu</code></h1>
+<p>A collection of GPU Widget classes</p>
+<p></p>
+
+
+<h2><a href="#Class_Gpu">Class Gpu </a></h2>
+<table class="function_list">
+	<tr>
+	<td class="name" nowrap><a href="#Gpu:init">Gpu:init()</a></td>
+	<td class="summary">no options</td>
+	</tr>
+</table>
+<h2><a href="#Class_GpuTop">Class GpuTop </a></h2>
+<table class="function_list">
+	<tr>
+	<td class="name" nowrap><a href="#GpuTop:init">GpuTop:init(args)</a></td>
+	<td class="summary"></td>
+	</tr>
+</table>
+
+<br/>
+<br/>
+
+
+    <h2 class="section-header has-description"><a name="Class_Gpu"></a>Class Gpu </h2>
+
+          <div class="section-description">
+          Compound widget to display GPU and VRAM usage.
+          </div>
+    <dl class="function">
+    <dt>
+    <a name = "Gpu:init"></a>
+    <strong>Gpu:init()</strong>
+    </dt>
+    <dd>
+    no options
+
+
+
+
+
+
+
+</dd>
+</dl>
+    <h2 class="section-header has-description"><a name="Class_GpuTop"></a>Class GpuTop </h2>
+
+          <div class="section-description">
+          Table of processes for the GPU, sorted by VRAM usage
+          </div>
+    <dl class="function">
+    <dt>
+    <a name = "GpuTop:init"></a>
+    <strong>GpuTop:init(args)</strong>
+    </dt>
+    <dd>
+
+
+
+    <h3>Parameters:</h3>
+    <ul>
+        <li><span class="parameter">args</span> table of options
+        <ul>
+        <li><span class="parameter">lines</span>
+            <span class="types">optional <span class="type">int</span></span>
+         how many processes to display
+         (<em>default</em> 5)
+        </li>
+        <li><span class="parameter">font_family</span>
+            <span class="types">optional <span class="type">string</span></span>
+
+        </li>
+        <li><span class="parameter">font_size</span>
+            <span class="types">optional <span class="type">number</span></span>
+
+        </li>
+        <li><span class="parameter">color</span>
+            <span class="types">optional <span class="type">{number,number,number}</span></span>
+         (default: <code>default_text_color</code>)
+        </li>
+        </li></ul>
+    </ul>
+
+
+
+
+
+</dd>
+</dl>
+
+
+</div> <!-- id="content" -->
+</div> <!-- id="main" -->
+<div id="about">
+<i>generated by <a href="http://github.com/stevedonovan/LDoc">LDoc 1.4.6</a></i>
+<i style="float:right;">Last updated 2023-05-01 20:49:18 </i>
+</div> <!-- id="about" -->
+</div> <!-- id="container" -->
+</body>
+</html>
