+++ docs/modules/widget_text.html
@@ -1,0 +1,244 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"
+   "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
+<html>
+<meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
+<head>
+    <title>Reference</title>
+    <link rel="stylesheet" href="../ldoc.css" type="text/css" />
+</head>
+<body>
+
+<div id="container">
+
+<div id="product">
+	<div id="product_logo"></div>
+	<div id="product_name"><big><b></b></big></div>
+	<div id="product_description"></div>
+</div> <!-- id="product" -->
+
+
+<div id="main">
+
+
+<!-- Menu -->
+
+<div id="navigation">
+<br/>
+<h1>polycore</h1>
+
+<ul>
+  <li><a href="../index.html">Index</a></li>
+</ul>
+
+<h2>Contents</h2>
+<ul>
+<li><a href="#Class_Text">Class Text </a></li>
+<li><a href="#Class_StaticText">Class StaticText </a></li>
+<li><a href="#Class_TextLine">Class TextLine </a></li>
+</ul>
+
+
+<h2>Modules</h2>
+<ul class="nowrap">
+  <li><a href="../modules/cairo_helpers.html">cairo_helpers</a></li>
+  <li><a href="../modules/data.html">data</a></li>
+  <li><a href="../modules/polycore.html">polycore</a></li>
+  <li><a href="../modules/util.html">util</a></li>
+  <li><a href="../modules/widget_core.html">widget_core</a></li>
+  <li><a href="../modules/widget_cpu.html">widget_cpu</a></li>
+  <li><a href="../modules/widget_drive.html">widget_drive</a></li>
+  <li><a href="../modules/widget_gpu.html">widget_gpu</a></li>
+  <li><a href="../modules/widget_graph.html">widget_graph</a></li>
+  <li><a href="../modules/widget_memory.html">widget_memory</a></li>
+  <li><a href="../modules/widget_network.html">widget_network</a></li>
+  <li><strong>widget_text</strong></li>
+</ul>
+
+</div>
+
+<div id="content">
+
+<h1>Module <code>widget_text</code></h1>
+<p>A collection of Widget classes</p>
+<p></p>
+
+
+<h2><a href="#Class_Text">Class Text </a></h2>
+<table class="function_list">
+	<tr>
+	<td class="name" nowrap><a href="#Text:init">Text:init(args)</a></td>
+	<td class="summary"></td>
+	</tr>
+</table>
+<h2><a href="#Class_StaticText">Class StaticText </a></h2>
+<table class="function_list">
+	<tr>
+	<td class="name" nowrap><a href="#StaticText:init">StaticText:init(text, args)</a></td>
+	<td class="summary"></td>
+	</tr>
+</table>
+<h2><a href="#Class_TextLine">Class TextLine </a></h2>
+<table class="function_list">
+	<tr>
+	<td class="name" nowrap><a href="#TextLine:init">TextLine:init(args)</a></td>
+	<td class="summary"></td>
+	</tr>
+	<tr>
+	<td class="name" nowrap><a href="#TextLine:set_text">TextLine:set_text(text)</a></td>
+	<td class="summary">Update the text line to be displayed.</td>
+	</tr>
+</table>
+
+<br/>
+<br/>
+
+
+    <h2 class="section-header has-description"><a name="Class_Text"></a>Class Text </h2>
+
+          <div class="section-description">
+          Common (abstract) base class for <a href="../modules/widget_text.html#StaticText">StaticText</a> and <a href="../modules/widget_text.html#TextLine">TextLine</a>.
+          </div>
+    <dl class="function">
+    <dt>
+    <a name = "Text:init"></a>
+    <strong>Text:init(args)</strong>
+    </dt>
+    <dd>
+
+
+
+    <h3>Parameters:</h3>
+    <ul>
+        <li><span class="parameter">args</span> table of options
+        <ul>
+        <li><span class="parameter">align</span>
+            <span class="types">optional <span class="type">string</span></span>
+         "left" (default), "center" or "right"
+        </li>
+        <li><span class="parameter">font_family</span>
+            <span class="types">optional <span class="type">string</span></span>
+
+         (<em>default</em> w.default_font_family)
+        </li>
+        <li><span class="parameter">font_size</span>
+            <span class="types">optional <span class="type">number</span></span>
+
+         (<em>default</em> w.default_font_size)
+        </li>
+        <li><span class="parameter">font_slant</span>
+            <span class="types">optional <span class="type">cairo_font_slant_t</span></span>
+
+         (<em>default</em> CAIRO_FONT_SLANT_NORMAL)
+        </li>
+        <li><span class="parameter">font_weight</span>
+            <span class="types">optional <span class="type">cairo_font_weight_t</span></span>
+
+         (<em>default</em> CAIRO_FONT_WEIGHT_NORMAL)
+        </li>
+        <li><span class="parameter">color</span>
+            <span class="types">optional <span class="type">{number,number,number,number}</span></span>
+         (default: <code>default_text_color</code>)
+        </li>
+        </li></ul>
+    </ul>
+
+
+
+
+
+</dd>
+</dl>
+    <h2 class="section-header has-description"><a name="Class_StaticText"></a>Class StaticText </h2>
+
+          <div class="section-description">
+          Draw some unchangeable text.
+ Use this widget for text that will never be updated.Text
+          </div>
+    <dl class="function">
+    <dt>
+    <a name = "StaticText:init"></a>
+    <strong>StaticText:init(text, args)</strong>
+    </dt>
+    <dd>
+
+
+
+    <h3>Parameters:</h3>
+    <ul>
+        <li><span class="parameter">text</span>
+            <span class="types"><span class="type">string</span></span>
+         Text to be displayed.
+        </li>
+        <li><span class="parameter">args</span>
+            <span class="types">optional <span class="type">table</span></span>
+         table of options, see <a href="../modules/widget_text.html#Text:init">Text:init</a>
+        </li>
+    </ul>
+
+
+
+
+
+</dd>
+</dl>
+    <h2 class="section-header has-description"><a name="Class_TextLine"></a>Class TextLine </h2>
+
+          <div class="section-description">
+          Draw a single line of changeable text.
+ Text line can be updated on each cycle via <a href="../modules/widget_text.html#TextLine:set_text">set_text</a>.
+          </div>
+    <dl class="function">
+    <dt>
+    <a name = "TextLine:init"></a>
+    <strong>TextLine:init(args)</strong>
+    </dt>
+    <dd>
+
+
+
+    <h3>Parameters:</h3>
+    <ul>
+        <li><span class="parameter">args</span>
+            <span class="types"><span class="type">table</span></span>
+         table of options, see <a href="../modules/widget_text.html#Text:init">Text:init</a>
+        </li>
+    </ul>
+
+
+
+
+
+</dd>
+    <dt>
+    <a name = "TextLine:set_text"></a>
+    <strong>TextLine:set_text(text)</strong>
+    </dt>
+    <dd>
+    Update the text line to be displayed.
+
+
+    <h3>Parameters:</h3>
+    <ul>
+        <li><span class="parameter">text</span>
+            <span class="types"><span class="type">string</span></span>
+
+        </li>
+    </ul>
+
+
+
+
+
+</dd>
+</dl>
+
+
+</div> <!-- id="content" -->
+</div> <!-- id="main" -->
+<div id="about">
+<i>generated by <a href="http://github.com/stevedonovan/LDoc">LDoc 1.4.6</a></i>
+<i style="float:right;">Last updated 2023-05-01 20:49:18 </i>
+</div> <!-- id="about" -->
+</div> <!-- id="container" -->
+</body>
+</html>
