+++ examples/text.lua
@@ -4,7 +4,8 @@
 local script_dir = debug.getinfo(1, 'S').source:match("^@(.*/)") or "./"
 package.path = script_dir .. "../?.lua;" .. package.path
 
-local widget = require('src/widget')
+local core = require('src/widgets/core')
+local text = require('src/widgets/text')
 local polycore = require('src/polycore')
 
 local width = 400
@@ -19,14 +20,14 @@ cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est
 laborum.]]
 
 --- Called once on startup to initialize widgets.
--- @treturn widget.Renderer
+-- @treturn core.Renderer
 function polycore.setup()
     local widgets = {
         -- heading
-        widget.Frame(widget.StaticText("Text Demo", {
+        core.Frame(text.StaticText("Text Demo", {
                 font_size=20,
                 font_weight=CAIRO_FONT_WEIGHT_BOLD,
-                color=widget.default_graph_color,
+                color=core.default_graph_color,
             }), {
             margin={0, 0, 10},
             border_sides={"bottom"},
@@ -35,21 +36,21 @@ function polycore.setup()
         }),
 
         -- simple text
-        widget.StaticText"Hello World!",
-        widget.Filler{height=10},
-        widget.StaticText("How are you doing?", {align="right"}),
+        text.StaticText"Hello World!",
+        core.Filler{height=10},
+        text.StaticText("How are you doing?", {align="right"}),
 
-        widget.Filler(),
+        core.Filler(),
 
         -- paragraph with newlines
-        widget.StaticText(LOREM_IPSUM, {
+        text.StaticText(LOREM_IPSUM, {
             align="center",
             font_slant=CAIRO_FONT_SLANT_ITALIC,
         }),
     }
 
     -- news ticker style text line
-    local ticker = widget.TextLine{align="center"}
+    local ticker = text.TextLine{align="center"}
     local line_width = 80  -- arbitrary estiamte
     local lipsum = LOREM_IPSUM:gsub("\n", " ")
     lipsum = lipsum .. " " .. lipsum:sub(1, line_width)
@@ -57,15 +58,15 @@ function polycore.setup()
         local offset = update_count % #lipsum
         self:set_text(lipsum:sub(offset, offset + line_width))
     end
-    table.insert(widgets, widget.Filler())
-    table.insert(widgets, widget.Frame(ticker, {
+    table.insert(widgets, core.Filler())
+    table.insert(widgets, core.Frame(ticker, {
         border_sides={"top"},
         border_width=1,
         border_color={1, 1, 1, .5},
     }))
 
-    local root = widget.Frame(widget.Rows(widgets), {margin=10})
-    return widget.Renderer{root=root, width=width, height=height}
+    local root = core.Frame(core.Rows(widgets), {margin=10})
+    return core.Renderer{root=root, width=width, height=height}
 end
